requirements.txt
pillow
africastalking
python-dotenv